feat(service): display a list of service for the current namespace
The earlier work to bring the OpenShift exention in has been modified.
No longer used to login, it is only here as an example.

The structure has been greatly modified and broken down into modules.

Also all references to 'odo' have been changed to 'kn'. This was done
just to get it to run. There is no plan to replace odo with the kn cli.